InsertComment2 Now checks the date of one file only, Also added a comparator
I'm not sure how to validate the time right now.
Please email me if something is confusing.
I'm not sure how we want to approach this yet,
So I wrote this for brainstorming/possibly using it for the 3rd
deliverable. If my code is hard to read please tell me to fix something
like add comments.