Added a prototype insert comment class.
This class works well, but there is no exceptional java error control.
I used the first formula of the wikipedia's the greate circle distance
formula, but the distance is a bit large from the real distance.
I am not sure that I used double values instead of DDMMSS values.
I forgot the latitude and longitude, DDMMSS and real number...a little
annoying works...